# Conflicts:
#	src/main/java/com/pshc/util/controller/PostController.java
#	src/main/java/com/pshc/util/dto/CategoryRepository.java

병합 후 커밋

+++ src/main/java/com/pshc/util/dto/CategoryRepository.java
@@ -8,12 +8,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.pshc.util.model.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer>{
-//	@Transactional
-//	@Modifying
-//	@Query("UPDATE Category c SET c.name = ?1, c.autoupdate = ?2, c.enable = ?3"
-//			+ ", c.visible = ?4, c.content = ?5 WHERE c.id = ?6") 
-//	int setCategoryFor(@Param("name")String name, @Param("autoupdate") String autoUpdate
-//			,@Param("enable") String enable, @Param("visible") String visible
-//			,@Param("content") String content, @Param("id") Long id);
+public interface CategoryRepository extends JpaRepository<Category, Long>{
+	
 }
